Added support for individual books
In the Bible's specific folder, add anew folder called taggedX where the individual books reside.
For example
NIV
     >tagged
            - NIV tagged text
     >taggedX
            - 01-Tagged Gen
            - 02-Tagged Exo
            - 03-Tagged Num
.            ...
Settings now has a check box to enable individual book selection.
The File Menu now has a menu item to open an individual file selection dialog.

+++ src/BibleTaggingUtil/BibleVersions/TargetVersion.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Reflection;
+using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -38,28 +39,73 @@ namespace BibleTaggingUtil.BibleVersions
 
                     if (bible.Count > 0)
                     {
-                        // construct Updates fileName
-                        string taggedFolder = Path.GetDirectoryName(container.Config.TaggedBible);
+                        bool individual = Properties.TargetBibles.Default.IndividualBooks;
+                        string currentbook = Properties.TargetBibles.Default.CurrentBook;
+
+                        string taggedFolder = string.Empty;
+                        if (individual)
+                        {
+                            if (string.IsNullOrEmpty(currentbook))
+                            {
+                                throw new Exception("No current book");
+                                //string biblesFolder = Properties.TargetBibles.Default.TargetBiblesFolder;
+                                //string target = Properties.TargetBibles.Default.TargetBible;
+                                //taggedFolder = Path.Combine(biblesFolder, target);
+                                //taggedFolder = Path.Combine(taggedFolder, "taggedX");
+                            }
+                            else
+                            {
+                                taggedFolder = Path.GetDirectoryName(currentbook);
+                            }
+                        }
+                        else
+                        {
+                            // construct Updates fileName
+                            taggedFolder = Path.GetDirectoryName(container.Config.TaggedBible);
+                        }
                         string oldTaggedFolder = Path.Combine(taggedFolder, "OldTagged");
                         if (!Directory.Exists(oldTaggedFolder))
                             Directory.CreateDirectory(oldTaggedFolder);
 
-                        // move existing tagged files to the old folder
-                        String[] existingTagged = Directory.GetFiles(taggedFolder, "*.*");
-                        foreach (String existingTaggedItem in existingTagged)
+                        string updatesFilePath = string.Empty;
+                        if (individual)
                         {
-                            string fName = Path.GetFileName(existingTaggedItem);
-                            string src = Path.Combine(taggedFolder, fName);
+                            // move current book to old folder
+                            string src = currentbook;
+                            string fName = Path.GetFileName(currentbook);
                             string dst = Path.Combine(oldTaggedFolder, fName);
-                            if (System.IO.File.Exists(dst))
-                                System.IO.File.Delete(src);
-                            else
-                                System.IO.File.Move(src, dst);
-                        }
+                            File.Move(src, dst);
 
-                        string baseName = Path.GetFileNameWithoutExtension(container.Config.TaggedBible);
-                        string updatesFileName = string.Format("{0:s}_{1:s}.txt", baseName, DateTime.Now.ToString("yyyy_MM_dd_HH_mm"));
-                        using (StreamWriter outputFile = new StreamWriter(Path.Combine(taggedFolder, updatesFileName)))
+                            // construct the updated file name
+                            string bareFilename = Path.GetFileNameWithoutExtension(currentbook);
+                            int speratorIndex = bareFilename.LastIndexOf("___");
+                            if(speratorIndex > 0)
+                                bareFilename = bareFilename.Substring(0, speratorIndex);
+                            string updatesFileName = $"{bareFilename}___{DateTime.Now.ToString("yyyy_MM_dd_HH_mm")}.txt";
+                            updatesFilePath = Path.Combine(taggedFolder, updatesFileName);
+                            Properties.TargetBibles.Default.CurrentBook = updatesFilePath;
+                            Properties.TargetBibles.Default.Save();
+                        }
+                        else
+                        {
+                            // move existing tagged files to the old folder
+                            String[] existingTagged = Directory.GetFiles(taggedFolder, "*.*");
+                            foreach (String existingTaggedItem in existingTagged)
+                            {
+                                string fName = Path.GetFileName(existingTaggedItem);
+                                string src = Path.Combine(taggedFolder, fName);
+                                string dst = Path.Combine(oldTaggedFolder, fName);
+                                if (System.IO.File.Exists(dst))
+                                    System.IO.File.Delete(src);
+                                else
+                                    System.IO.File.Move(src, dst);
+                            }
+
+                            string baseName = Path.GetFileNameWithoutExtension(container.Config.TaggedBible);
+                            string updatesFileName = string.Format("{0:s}_{1:s}.txt", baseName, DateTime.Now.ToString("yyyy_MM_dd_HH_mm"));
+                            updatesFilePath = Path.Combine(taggedFolder, updatesFileName);
+                        }
+                        using (StreamWriter outputFile = new StreamWriter(updatesFilePath))
                         {
                             foreach (string verseRef in  container.Target.Bible.Keys)
                             {
